v1.3.3 update
Change ManualRx1 to Rx1Manual, etc. Altered dataReceived code to use
defined instructions names., and to only change button colours when an
instruction is received from the box (and removed this code from the
button click handlers). This validates that instructions are actually
received and acted on. Changed SendCommand(0); to
SendCommand(Constants.CheckConnection);